requirements.txt
pandas~=2.2.3
numpy
yfinance~=0.2.55
backtrader
tensorflow
scikit-learn
pymysql~=1.0.2
sqlalchemy
django-scheduler~=0.10.1